move test setup mutations out of the databasemutationbuilder
