+++ server/api/src/test/scala/com/prisma/api/ApiTestDatabase.scala
@@ -2,7 +2,7 @@ package com.prisma.api
 
 import akka.actor.ActorSystem
 import akka.stream.ActorMaterializer
-import com.prisma.api.database.{DatabaseMutationBuilder, DatabaseQueryBuilder}
+import com.prisma.api.database.{DatabaseApiTestDatabaseMutationBuilder, DatabaseQueryBuilder}
 import com.prisma.deploy.migration.mutactions.{ClientSqlMutaction, CreateRelationTable}
 import com.prisma.shared.models._
 import com.prisma.utils.await.AwaitUtils
@@ -27,7 +27,7 @@ case class ApiTestDatabase()(implicit dependencies: ApiDependencies) extends Awa
   def truncate(project: Project): Unit = {
     val tables = clientDatabase.run(DatabaseQueryBuilder.getTables(project.id)).await
     val dbAction = {
-      val actions = List(sqlu"""USE `#${project.id}`;""") ++ List(DatabaseMutationBuilder.dangerouslyTruncateTable(tables))
+      val actions = List(sqlu"""USE `#${project.id}`;""") ++ List(DatabaseApiTestDatabaseMutationBuilder.dangerouslyTruncateTable(tables))
       DBIO.seq(actions: _*)
     }
     clientDatabase.run(dbAction).await()
@@ -35,8 +35,9 @@ case class ApiTestDatabase()(implicit dependencies: ApiDependencies) extends Awa
 
   def delete(project: Project): Unit = dropDatabases(Vector(project.id))
 
-  private def createProjectDatabase(project: Project)                   = runDbActionOnClientDb(DatabaseMutationBuilder.createClientDatabaseForProject(project.id))
-  private def createModelTable(project: Project, model: Model)          = runDbActionOnClientDb(DatabaseMutationBuilder.createTableForModel(project.id, model))
+  private def createProjectDatabase(project: Project) = runDbActionOnClientDb(DatabaseApiTestDatabaseMutationBuilder.createClientDatabaseForProject(project.id))
+  private def createModelTable(project: Project, model: Model) =
+    runDbActionOnClientDb(DatabaseApiTestDatabaseMutationBuilder.createTableForModel(project.id, model))
   private def createRelationTable(project: Project, relation: Relation) = runMutaction(CreateRelationTable(project.id, project.schema, relation = relation))
 
   def deleteExistingDatabases(): Unit = {
@@ -50,7 +51,7 @@ case class ApiTestDatabase()(implicit dependencies: ApiDependencies) extends Awa
   }
 
   private def dropDatabases(dbs: Vector[String]): Unit = {
-    val dbAction = DBIO.seq(dbs.map(db => DatabaseMutationBuilder.dropDatabaseIfExists(database = db)): _*)
+    val dbAction = DBIO.seq(dbs.map(db => DatabaseApiTestDatabaseMutationBuilder.dropDatabaseIfExists(database = db)): _*)
     clientDatabase.run(dbAction).await(60)
   }
